Move class builder methods so zeitwerk can find them
